% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rDirichlet.R
\name{rDirichlet}
\alias{rDirichlet}
\title{Random Dirichlet Draws}
\usage{
rDirichlet(n, alpha)
}
\arguments{
\item{n}{number of draws}

\item{alpha}{parameter vector}
}
\value{
what the function returns
}
\description{
Draws from Dirichlet distribution
}
\author{
Federico Mattiello <Federico.Mattiello@UGent.be>
}

