% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrapperWMW.R
\name{wrapperWMW}
\alias{wrapperWMW}
\title{Wrapper function for Wilcoxon-Mann-Whitney test}
\usage{
wrapperWMW(x, y, adjMethod = "BH", ...)
}
\arguments{
\item{x}{\code{matrix} containing the first sample}

\item{y}{\code{matrix} containing the second sample}

\item{...}{other parameters for \link{\code{wilcox.test}}}
}
\value{
\code{numeric} vector of length \code{ncol(x)} containing adjusted p.values
}
\description{
Wrapper function for Wilcoxon-Mann-Whitney test
}
\details{
Utility function performing WMW test on each OTU individually, and then correcting
for multiplicity with one of the standard corrections, \code{"fdr"} is the default.
}
\author{
Federico Mattiello <Federico.Mattiello@UGent.be>
}

